+++ samengine-build/src/cli/argparser.ts
@@ -4,7 +4,6 @@
 export interface CLIArgs {
     release: boolean;
     singlefile: boolean;
-    port: number;
     newProject: string | null;
     empty: boolean;
 }
@@ -12,7 +11,7 @@ export interface CLIArgs {
 // Function to parse the Args for the CLI Tools
 export function parseArgs(): CLIArgs {
     const args = process.argv.slice(2);
-    const options: CLIArgs = { release: false, singlefile: false, port: 3000, newProject: null, empty: false };
+    const options: CLIArgs = { release: false, singlefile: false, newProject: null, empty: false };
 
     for (let i = 0; i < args.length; i++) {
         const arg = args[i];
@@ -24,10 +23,6 @@ export function parseArgs(): CLIArgs {
             case "-r":
                 options.release = true;
                 break;
-            case "--port":
-            case "-p":
-                options.port = Number(args[++i]);
-                break;
             case "--new":
             case "-n":
                 options.newProject = args[++i];
@@ -38,7 +33,7 @@ export function parseArgs(): CLIArgs {
                 break;
             case "-h":
             case "--help":
-                console.log("CLI Tools for samengine\nUsage:\n  -r, --release\n  -p <port>\n  -n <project>\n  --new-empty\n --single-file   to generate the Export into one file");
+                console.log("CLI Tools for samengine\nUsage:\n  -r, --release\n  -n <project>\n  --new-empty\n --single-file   to generate the Export into one file");
                 process.exit(0);
             default:
                 console.warn(`⚠️ Unknown Argument: ${arg}`);