+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 ## samengine-build (1.9.2) (22.04.2026)
 
 - fixed a Bug with the Outdir
+- dev Server Port can now be set in the Project Config File
+- the server will add the Port number +1 when the current set
+Dev Port is busy instead of crashing.
 
 
 ## samengine-build (1.9.1) (21.04.2026)